% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auc_trapz}
\alias{auc_trapz}
\title{Area Under the Curve (AUC) via Trapezoidal Rule}
\usage{
auc_trapz(TPR, FPR)
}
\arguments{
\item{TPR}{Numeric vector of true positive rates (y-axis)}

\item{FPR}{Numeric vector of false positive rates (x-axis)}
}
\value{
Numeric value of AUC
}
\description{
Computes the area under a curve (AUC) using the trapezoidal rule. Used for ROC curves.
}
\keyword{internal}
