% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuants.R
\name{compareTS,FLQuant,FLQuant-method}
\alias{compareTS,FLQuant,FLQuant-method}
\title{Time Series Comparison for FLQuant}
\usage{
\S4method{compareTS}{FLQuant,FLQuant}(obs, pred)
}
\arguments{
\item{obs}{FLQuant object of observed time series}

\item{pred}{FLQuant object of predicted time series}
}
\value{
Data.frame with similarity metrics
}
\description{
Calculate similarity measures for FLQuant time series
}
