% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{compareTS}
\alias{compareTS}
\alias{compareTS,numeric,numeric-method}
\title{Time Series Comparison}
\usage{
compareTS(obs, pred)

\S4method{compareTS}{numeric,numeric}(obs, pred)
}
\arguments{
\item{obs}{Numeric vector of observed time series}

\item{pred}{Numeric vector of predicted time series}
}
\value{
Data.frame with similarity metrics
}
\description{
Calculate similarity measures between time series
}
