% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/roc.R, R/FLQuants.R
\name{PN}
\alias{PN}
\alias{PN,numeric,numeric-method}
\alias{PN,FLQuant,FLQuant-method}
\title{Confusion Matrix Statistics}
\usage{
PN(response, predictor)

\S4method{PN}{numeric,numeric}(response, predictor)

\S4method{PN}{FLQuant,FLQuant}(response, predictor)
}
\arguments{
\item{response}{Numeric vector of observed values}

\item{predictor}{Numeric vector of predicted values}
}
\value{
A data frame containing TP, TN, FP, and FN counts
}
\description{
Calculates True Positive (TP), True Negative (TN), False Positive (FP), 
and False Negative (FN) counts from predicted and actual values
}
\examples{
\dontrun{
response = c(1, -1, 1, -1)
predictor = c(1, -1, -1, 1)
confusionStats = PN(response, predictor)
}

}
