% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/FLQuants.R
\name{roc2}
\alias{roc2}
\alias{roc2,FLQuant,FLQuant-method}
\title{Calculate ROC (Receiver Operating Characteristic) statistics for two numeric vectors.}
\usage{
roc2(response, predictor, ...)

\S4method{roc2}{FLQuant,FLQuant}(response, predictor, ...)
}
\arguments{
\item{response}{A numeric vector representing the response values.}

\item{predictor}{A numeric vector representing the predictor values.}
}
\value{
A data frame containing the following columns:
  \describe{
    \item{response}{The response values.}
    \item{label}{A logical vector indicating whether each response is greater than 1 (TRUE) or not (FALSE).}
    \item{predictor}{The predictor values.}
    \item{TPR}{The True Positive Rate (TPR) calculated as TP / (TP + FN).}
    \item{FPR}{The False Positive Rate (FPR) calculated as FP / (FP + TN).}
    \item{TP}{The True Positives (TP).}
    \item{TN}{The True Negatives (TN).}
    \item{FP}{The False Positives (FP).}
    \item{FN}{The False Negatives (FN).}
    \item{TSS}{The True Skill Score (TSS) calculated as (TP / (TP + FN)) - (FP / (FP + TN)).}
    \item{order}{The order of the predictor values after sorting in descending order.}
  }
}
\description{
This function calculates ROC statistics, including True Positive Rate (TPR), False Positive Rate (FPR), True Positives (TP), True Negatives (TN), False Positives (FP), False Negatives (FN), and True Skill Score (TSS) for two numeric vectors.
}
