% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuants.R
\name{state,FLQuant,FLQuant-method}
\alias{state,FLQuant,FLQuant-method}
\title{State Classification for FLQuant}
\usage{
\S4method{state}{FLQuant,FLQuant}(obs, pred)
}
\arguments{
\item{obs}{FLQuant object of observed stock status}

\item{pred}{FLQuant object of predicted stock status}
}
\value{
Data.frame with classification metrics
}
\description{
Evaluate stock status classification for FLQuant objects
}
