% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuants.R
\name{trend,FLQuant,FLQuant-method}
\alias{trend,FLQuant,FLQuant-method}
\title{Trend Analysis for FLQuant}
\usage{
\S4method{trend}{FLQuant,FLQuant}(obs, pred)
}
\arguments{
\item{obs}{FLQuant object of observed time series}

\item{pred}{FLQuant object of predicted time series}
}
\value{
Data.frame with trend metrics
}
\description{
Calculate trend agreement metrics for FLQuant objects
}
