% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{trend}
\alias{trend}
\alias{trend,numeric,numeric-method}
\title{Trend Analysis}
\usage{
trend(obs, pred)

\S4method{trend}{numeric,numeric}(obs, pred)
}
\arguments{
\item{obs}{Numeric vector of observed time series}

\item{pred}{Numeric vector of predicted time series}
}
\value{
Data.frame with trend metrics
}
\description{
Calculate trend agreement metrics between time series
}
