% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{variability}
\alias{variability}
\alias{variability,numeric,numeric-method}
\title{Variability Comparison}
\usage{
variability(response, predictor)

\S4method{variability}{numeric,numeric}(response, predictor)
}
\arguments{
\item{response}{Numeric vector of observed values}

\item{predictor}{Numeric vector of predicted values}
}
\value{
Data.frame with variability ratios
}
\description{
Compare variability characteristics between time series
}
