% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuants.R
\name{ccfFn,FLQuant,FLQuant-method}
\alias{ccfFn,FLQuant,FLQuant-method}
\title{Cross-Correlation Function for FLQuant}
\usage{
\S4method{ccfFn}{FLQuant,FLQuant}(obs, pred, lag.max = 5)
}
\arguments{
\item{obs}{FLQuant object of observed time series}

\item{pred}{FLQuant object of predicted time series}

\item{lag.max}{Maximum lag to consider (default=5)}
}
\value{
Data.frame with optimal lag and ACF value
}
\description{
Find optimal lag between FLQuant time series
}
