% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{ccfFn}
\alias{ccfFn}
\alias{ccfFn,numeric,numeric-method}
\title{Cross-Correlation Function}
\usage{
ccfFn(response, predictor, lagMax = 5)

\S4method{ccfFn}{numeric,numeric}(response, predictor, lagMax = 5)
}
\arguments{
\item{response}{Numeric vector of observed time series}

\item{predictor}{Numeric vector of predicted time series}

\item{lagMax}{Maximum lag to consider}
}
\value{
Data.frame with optimal lag and ACF value
}
\description{
Find optimal lag between time series
}
