% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{ccfFn}
\alias{ccfFn}
\alias{ccfFn,numeric,numeric-method}
\title{Cross-Correlation Function}
\usage{
ccfFn(obs, pred, lag.max = 5)

\S4method{ccfFn}{numeric,numeric}(obs, pred, lag.max = 5)
}
\arguments{
\item{obs}{Numeric vector of observed time series}

\item{pred}{Numeric vector of predicted time series}

\item{lag.max}{Maximum lag to consider}
}
\value{
Data.frame with optimal lag and ACF value
}
\description{
Find optimal lag between time series
}
