% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conf_matrix_basic}
\alias{conf_matrix_basic}
\title{Basic Confusion Matrix}
\usage{
conf_matrix_basic(response, predictor)
}
\arguments{
\item{response}{Logical or binary vector of observed values}

\item{predictor}{Logical or binary vector of predicted values}
}
\value{
Data.frame with TP, TN, FP, FN
}
\description{
Computes basic confusion matrix statistics for binary classification.
}
\keyword{internal}
