% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/roc.R
\name{PN}
\alias{PN}
\alias{PN,numeric,numeric-method}
\title{Confusion Matrix Statistics}
\usage{
PN(obs, hat)

\S4method{PN}{numeric,numeric}(obs, hat)
}
\arguments{
\item{obs}{Numeric vector of predicted values}

\item{hat}{Numeric vector of actual values}
}
\value{
A data frame containing TP, TN, FP, and FN counts
}
\description{
Calculates True Positive (TP), True Negative (TN), False Positive (FP), 
and False Negative (FN) counts from predicted and actual values
}
\examples{
\dontrun{
pred <- c(1, -1, 1, -1)
actual <- c(1, -1, -1, 1)
confusion_stats <- PN(pred, actual)
}

}
