% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/roc.R, R/FLQuants.R
\name{PN}
\alias{PN}
\alias{PN,numeric,numeric-method}
\alias{PN,FLQuant,FLQuant-method}
\title{Confusion Matrix Statistics}
\usage{
PN(obs, pred)

\S4method{PN}{numeric,numeric}(obs, pred)

\S4method{PN}{FLQuant,FLQuant}(obs, pred)
}
\arguments{
\item{obs}{Numeric vector of observed values}

\item{pred}{Numeric vector of predicted values}
}
\value{
A data frame containing TP, TN, FP, and FN counts
}
\description{
Calculates True Positive (TP), True Negative (TN), False Positive (FP), 
and False Negative (FN) counts from predicted and actual values
}
\examples{
\dontrun{
obs = c(1, -1, 1, -1)
pred = c(1, -1, -1, 1)
confusionStats = PN(obs, pred)
}

}
