% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuants.R
\name{skillScore,FLQuant,FLQuant-method}
\alias{skillScore,FLQuant,FLQuant-method}
\title{Skill Score for FLQuant}
\usage{
\S4method{skillScore}{FLQuant,FLQuant}(response, predictor, reference = NULL, threshold = 1)
}
\arguments{
\item{response}{FLQuant object of observed values}

\item{predictor}{FLQuant object of predicted values}

\item{reference}{Reference value for classification (default=1)}

\item{threshold}{Threshold for classification (default=1)}
}
\value{
Data.frame with skill metrics
}
\description{
Calculate skill scores for FLQuant objects
}
