% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{state}
\alias{state}
\alias{state,numeric,numeric-method}
\title{State Classification}
\usage{
state(response, predictor)

\S4method{state}{numeric,numeric}(response, predictor)
}
\arguments{
\item{response}{Numeric vector of observed stock status}

\item{predictor}{Numeric vector of predicted stock status}
}
\value{
Data.frame with classification metrics
}
\description{
Evaluate stock status classification performance
}
