% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{trend}
\alias{trend}
\alias{trend,numeric,numeric-method}
\title{Trend Analysis}
\usage{
trend(response, predictor)

\S4method{trend}{numeric,numeric}(response, predictor)
}
\arguments{
\item{response}{Numeric vector of observed time series}

\item{predictor}{Numeric vector of predicted time series}
}
\value{
Data.frame with trend metrics
}
\description{
Calculate trend agreement metrics between time series
}
