% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/roc.R
\name{TSS}
\alias{TSS}
\alias{TSS,numeric,numeric,numeric,numeric-method}
\title{Calculate True Skill Statistic (TSS)}
\usage{
TSS(TP, TN, FP, FN)

\S4method{TSS}{numeric,numeric,numeric,numeric}(TP, TN, FP, FN)
}
\arguments{
\item{TP}{Number of True Positives}

\item{TN}{Number of True Negatives}

\item{FP}{Number of False Positives}

\item{FN}{Number of False Negatives}
}
\value{
Numeric value of TSS (sensitivity + specificity - 1)
}
\description{
Calculates the True Skill Statistic (TSS) from confusion matrix elements
}
\details{
TSS ranges from -1 to +1, where +1 indicates perfect agreement and values of zero 
or less indicate a performance no better than random
}
\examples{
\dontrun{
tss_value <- TSS(TP=10, TN=8, FP=2, FN=3)
}

}
