% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuants.R
\name{variability,FLQuant,FLQuant-method}
\alias{variability,FLQuant,FLQuant-method}
\title{Variability Comparison for FLQuant}
\usage{
\S4method{variability}{FLQuant,FLQuant}(response, predictor)
}
\arguments{
\item{response}{FLQuant object of observed values}

\item{predictor}{FLQuant object of predicted values}
}
\value{
Data.frame with variability ratios
}
\description{
Compare variability characteristics for FLQuant objects
}
